+++ ChangeLog.md
@@ -13,8 +13,8 @@ and this project adheres to
 
 ### Added
 
-- a DAG (direct acyclic graph)  with topological sorting to the Venice 
-  data types. E.g.: `(topological-sort (dag ["A" "B"] ["B" "C"]))`
+- a DAG (direct acyclic graph) with topological sorting.
+  E.g.: `(topological-sort (dag ["A" "B"] ["B" "C"]))`